Refactor throws.

+++ base/collection/HashSet.h
@@ -434,11 +434,14 @@ public:
     /**
       Removes the specified value from the set.
     */
-    void remove(const Value& value) throw(InvalidNode) {
+    void remove(const Value& value) throw(InvalidNode)
+    {
       const unsigned long hash = getHash(value);
       Node** bucket = getBuckets() + (hash & mask);
       Node* child = *bucket;
-      bassert(child, InvalidNode(this));
+      if (!child) {
+        throw InvalidNode(this);
+      }
       if ((child->getHash() == hash) && (child->getValue() == value)) {
         *bucket = child->getNext(); // unlink first node (next could be 0)
       } else {
@@ -448,7 +451,9 @@ public:
           parent = child;
           child = child->getNext();
         }
-        bassert(child, InvalidNode(this));
+        if (!child) {
+          throw InvalidNode(this);
+        }
         parent->setNext(child->getNext()); // unlink node from linked list
       }
       --size;
@@ -530,7 +535,9 @@ public:
       EndOfEnumeration if the end has been reached.
     */
     Pointer next() throw(EndOfEnumeration) {
-      bassert(numberOfElements, EndOfEnumeration(this));
+      if (!numberOfElements) {
+        throw EndOfEnumeration(this);
+      }
       while (!node) {
         ++bucket;
         node = *bucket;
